Add files via upload
The model presented here is an RNN model that is used for predicting google stock prices. The accuracy is quite high. It will also give beginners an insight on to how to go about building a basic RNN model. The dataset that is given here has five years of data. The model can predict the stock market values only for the next day.